Added config.toml with themes

+++ controller/controller.py
@@ -47,13 +47,15 @@ class Controller:
 
     def load_image_callback(self):
         """Open file dialog and load the selected image into the view."""
-        self._image_path = xdialog.open_file(
+        path_tmp = xdialog.open_file(
             title="Select Image",
             filetypes=[("Images", "*.png *.jpg *.jpeg")],
             multiple=False
         )
-        dpg.delete_item(item=Tag.CURRENT_IMAGE_TEXTURE, children_only=False)
-        dpg.delete_item(item=Tag.CURRENT_IMAGE, children_only=False)
+        if path_tmp:
+            self._image_path = path_tmp
+            dpg.delete_item(item=Tag.CURRENT_IMAGE_TEXTURE, children_only=False)
+            dpg.delete_item(item=Tag.CURRENT_IMAGE, children_only=False)
 
         if self._image_path:
             width, height, _, data = dpg.load_image(self._image_path)
